chore: move fake ui bridge to utils
